+++ changelog.md
@@ -3,6 +3,8 @@
 ## Upcoming
 Requires InsaneLib 1.21.12 and forge 47.4.0  
 **Config options might have been reset**
+* Stamina consumption is now reduced by 20% when out of combat
+  * Out of combat = not attacked or hurt in the last 15 seconds
 * Fixed empty stamina on respawn
 
 ## Beta 1.3.1